+++ src/home_work_3/calcs/simple/CalculatorWithMathExtends.java
@@ -59,8 +59,8 @@ public class CalculatorWithMathExtends extends  CalculatorWithOperator{
      * @return
      */
     @Override
-    public int modul (int a) {
-        int mod = Math.abs(a);
+    public double modul (double a) {
+        double mod = Math.abs(a);
         return mod;
     }
     /**
